requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
ortools==9.7.2996
